+++ types/seatsio/tsconfig.json
@@ -17,6 +17,11 @@
         "forceConsistentCasingInFileNames": true
     },
     "files": [
+        "lib/definitions.ts",
+        "lib/seatingChart.ts",
+        "lib/eventManager.ts",
+        "lib/chartManager.ts",
+        "lib/seatingChartDesigner.ts",
         "index.d.ts",
         "seatsio-tests.ts"
     ]
